AC-1902: Remove PHPCSUtils dependency from coding standard
