+++ Magento2/Helpers/PHPCSUtils/BackCompat/Helper.php
@@ -8,7 +8,7 @@
  * @link      https://github.com/PHPCSStandards/PHPCSUtils
  */
 
-namespace PHPCSUtils\BackCompat;
+namespace Magento2\Helpers\PHPCSUtils\BackCompat;
 
 use PHP_CodeSniffer\Exceptions\RuntimeException;
 use PHP_CodeSniffer\Files\File;
